KIWI-1684: Implement ApiUtils
