+++ src/tests/api/HappyPath.test.ts
@@ -1,10 +1,6 @@
-import { constants } from "../utils/ApiConstants";
-import { getKeyFromSession,
-	startStubServiceAndReturnSessionId,
-	validateTxMAEventData,
-	getTxmaEventsFromTestHarness,
-	wellKnownGet,
-} from "../utils/ApiTestSteps";
+import { constants } from "./ApiConstants";
+import { getKeyFromSession, startStubServiceAndReturnSessionId, wellKnownGet } from "./ApiTestSteps";
+import { getTxmaEventsFromTestHarness, validateTxMAEventData } from "./ApiUtils";
 
 describe("Happy path tests", () => {
 	describe("/session endpoint", () => {
@@ -15,21 +11,20 @@ describe("Happy path tests", () => {
 			const sessionResponse = await startStubServiceAndReturnSessionId(journeyType);
 			const sessionId = sessionResponse.data.session_id;
 
-			await expect(getKeyFromSession(sessionId, constants.DEV_CIC_SESSION_TABLE_NAME, "journey")).resolves.toBe(journeyType);
+			const savedJourney = getKeyFromSession(sessionId, constants.DEV_CIC_SESSION_TABLE_NAME, "journey");
+			await expect(savedJourney).resolves.toBe(journeyType);
 
 			const allTxmaEventBodies = await getTxmaEventsFromTestHarness(sessionId, 1);
 			validateTxMAEventData({ eventName: "CIC_CRI_START", schemaName }, allTxmaEventBodies);
 		});
 	});
 
-	describe("./wellknown/jwks.json endpoint", () => {
-		it("E2E Happy Path Journey - Well Known", async () => {
-			const { status, data } = await wellKnownGet();
+	it("./wellknown/jwks.json endpoint", async () => {
+		const { status, data } = await wellKnownGet();
 			
-			expect(status).toBe(200);
-			expect(data.keys).toHaveLength(2);
-			expect(data.keys[0].use).toBe("sig");
-			expect(data.keys[1].use).toBe("enc");
-		});
+		expect(status).toBe(200);
+		expect(data.keys).toHaveLength(2);
+		expect(data.keys[0].use).toBe("sig");
+		expect(data.keys[1].use).toBe("enc");
 	});
 });
